+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `firstName`, `lastName`, `fullName`, `email`, `phone` on dataLayer for userData event
 
 ## [3.0.2] - 2021-10-11
 ### Added